Nix the building of the getdown-tools.jar. We can just include everything in getdown-pro.jar and it's only 10% larger (109k -> 121k), and it vastly simplifies life for users.
Just use getdown-pro.jar for all of your Getdown-related needs:

- use it to install your application
- use it in your build scripts for the various Getdown tasks
- link against it in your app to use LaunchUtil or to embed Getdown

We'll even ship this on the website as simply getdown.jar. I'm tempted to ship
this in Maven and modify the main Getdown POM to not export any of its
dependencies. However, that will require a bunch of build file tweaking, which
I'd rather save for another day.