A hopefully kosher edit.
This little chunk of code has been tweaked and cargo'd forward for years
with no documentation as to what's going on.

What I object to is the blatent disregarding of the width specified in
getdown.txt with something calculated internally. (width - x*2)

GIGO; fix your fucking getdown.txt.

- If we are not granted privileges, reading the metadata breaks, which
is why the fallback was first added.
- Later it was tweaked because getWidth() can sometimes report 0 and
that would cause the fallback width to be negative.