Revert "Ant build's been deprecated for ages."
This reverts commit 12f7b666af164c7bca0bfa27756c1d4b86d352aa.

So it turns out despite getdown being public and having an
automatic travis build we still have an internal build. Which
uses ant. And requires blood sacrifices to cope with maven.

I'm really not in the mood to deal with it/don't have access to
set up the travis build to publish snapshots, so I'm just going to
revert and let someone else with more motivation clean up this mess
(or not)