Fix the getdown manifest for maven builds.
We have an ant build.xml and a maven pom.xml.  At some point I guess the
build switched to maven, which didn't have the special manifest to use
for the jar to work around all the Java permission changes.  If you want
getdown to work with the latest Java, you will need to use this update.