Added support for "strict comments".
In this mode, Getdown only treats a # as a comment starter in column zero. All
other #s are ignored. You enable it by putting:

strict_comments = true

in your getdown.txt file. Yes the same file that has to be parsed with or
without strict comments. If we see that, we reparse the file with strict
comments.