set window to decorated even if error occurs during very early stage
When the window is undecorated and an error occurs very early, e.g. during
java version check, the order of the calls in the fail method is
significant. super.fail() will call setStatusAsync, which will create
the interface using EventQueue.invokeLater, if it does not exist yet.

Therefore we must first call super.fail(), and then asynchronously make
the window decorated again. Otherwise the window might not be created yet.

Closes #57