Prevent Java VM update on Windows if vm is still in use
and log it accordingly.
This happens for instance when a user has the application still open
(maybe minimized), launches getdown and getdown tries to install a new
java vm. Currently in such situations the unpack does it's work until it
fails to open a file for overriding (typically awt.dll or java.dll).
This leaves the vm in a mixed state which is not satisfactory.
This patch verifies if java.dll is writeable before proceeding to update
the vm. 
I think this should also address issue #104