Move javadoc plugin to plugins instead of pluginManagement.
For some incomprehensible reason, Maven is not using the config for
maven-javadoc-plugin when I put it in pluginManagement (it reverts to version
2.7). But it does use the updated config for the maven-compiler-plugin and the
maven-resources-plugin. And even more bizarrely, it does use the
maven-javadoc-plugin config in plugin for all submodules even though the *whole
point* of the pluginManagement section is for config that applies to all
submodules. The plugin section *should not* apply to all submodules.

Even after all these years, Maven can still surprise me with new failures
stemming from its legion of unfortunate design decisions.