Use Graal JavaScript VM.
Nashorn was removed in JDK15, but the PAC proxy detection code depends on
having a JavaScript VM (what a twisty maze of passages). Rather than remove it
completely, I'll just keep the tests working so that people who use Getdown
with PAC proxies on JDK 14 or earlier can keep on trucking.

Perhaps the JavaScript code could be rewritten in Java and this weird massive
dependency could be eliminated?