JarDiffPatcher needs to support .old files.
When patching, Getdown will move foo.jar to foo.jar.old and then patch that
file to create a new foo.jar. That results in foo.jar.old being passed into
JarDiffPatcher as the source file, which it then inspects to determine whether
it is a zip file or a jar file based on its file extension. Since that file
extension has already changed to .xxx.old, we need to handle that.

Closes #230.