Upgrade Proguard to a version that supports Java 18.
For unfortunate reasons, we have to add a separate (build-time) dependency on
proguard-core, code from which proguard-base relies on but neither includes
directly nor expresses as a dependency.

How is it in the year of our lord two thousand and twenty two, we are still
making such basic build errors in our Maven plugins and POM files?